// src/app/modules/Patient/patient.interface.ts
import { BloodGroup, Gender, MaritalStatus } from "@prisma/client";

export type IPatientFilterRequest = {
  name?: string | undefined;
  email?: string | undefined;
  contactNumber: string | undefined;
  searchTerm: string | undefined;
};

type IPatientHealthData = {
  gender: Gender;
  dateOfBirth: string;
  bloodGroup: BloodGroup;
  hasAllergies?: boolean;
  hasDiabetes?: boolean;
  height: string;
  weight: string;
  smokingStatus?: boolean;
  dietaryPreferensec?: string;
  pregnancyStatus?: boolean;
  mentalHealthHistory?: string;
  immunizationStatus?: string;
  hasPastSurgeries?: boolean;
  recentAnxiety?: boolean;
  recentDepression?: boolean;
  maritalStatus?: MaritalStatus;
};

type IMedicalReport = {
  reportName: string;
  reportLink: string;
};

export type IPatientUpdte = {
  name: string;
  contactNumber: string;
  address: string;
  patientHealthData?: IPatientHealthData;
  medicalReport?: IMedicalReport;
};